Pattern: Add array
Allow let statements of the form

let [a, b, c] = [1, 2, 3];

Internally, this is translated into

let ((a, b), c) = ((1, 2), 3)

Array patters are balanced jbinary trees of patterns,
just like arrays are balanced binary trees of values.

Computing a program that returns the value of an identifier inside a
pattern becomes more complicated with arrays. We use a combination of
BinaryTree and DirectedTree to track the path of the identifier from the
pattern root.