Pattern: Convert to base
Array types and values are macros that don't exist in the underlying
Simplicity. [1, 2, 3] is equivalent to ((1, 2), 3). In the next commit
we will introduce array patterns of the form let [a, b, c] = [1, 2, 3].
This is equivalent to let ((a, b), c) = ((1, 2), 3). We this translated
pattern a _base pattern_ because it corresponds to the underlying
Simplcity. As we add more types and patterns, it will be useful to rely
on base patterns, to prevent too many edge cases.