Compile with errors
Sanket designed the ProgExt trait without errors because the Simfony
compiler is supposed to construct valid Simplicity in every step.
The compiler needs a type system to be this careful, which is not
implemented yet. We must propagate the low-level simplicity errors.

This is why I add errors to the fallible combinators in the ProgExt
trait. Once Simfony has a type system, we should revert this change.