Compile with errors
Replay pretty compile errors to the user.
Handle arrays / list expressions in the next commit and introduce
temporary unwraps. These are the only unwraps introduced by this commit.

During the translation from Simfony to Simplicity, the only errors we
should get are from the comp and case combinator.

assert{l,r} cannot fail because there is only one branch. The upstream
methods return Result, which is wrong. I will push a fixup to upstream.

pair should never fail because during the translation of Simfony to
Simplicity, the source type is variable. This is hard to prove, however,
so I chose to handle potential pair unification errors. In a future PR,
I hope to write a proof and remove all these nonexistent error paths.