Compile (arrays|lists) with errors
Arrays and lists are nasty:
The compiler constructs a lot of products which should never fail to
unify... but I cannot convincingly prove it, so I must handle every
single pair as potentially failing... In the end, I fold over tree of
Result<ProgNode, _> instead of trees of ProgNode. This blows up the
compiler code; sorry. I think it is still readable, but we should work
on those infalliable pairs :)