Introduce infallible assertions
Because assert{l,r} has only one child, the target type of the left child
is never unified with the source type of the right child. By using
`take s` or `drop s` as the child, the source type is always a produce
type.

It turns out, assert{l,r} CAN fail if the child has a unit or sum source
type. This is the only possible error path.