Scope: Remove implicit unit value
GlobalScope assumed a unit value inside the input value because the
translation of Simfony to Simplicity includes such a unit value.
However, we will create scopes that track values without unit.

A pattern for input values is required to construct a scope. We remove
the default scope constructor because there is no default pattern.