Remove witness tracker
Each witness node has a name, which is matched against a JSON file to
assign its value. The scope is not needed for this. If we want to
enumerate all required witness names, then we can walk through the
compiled Simplicity program.

Witness names are unscoped but the variables that hold them are scoped.
witness("name") returns a constant value like a jet jet_version.