Add Value + ValueConstructible
Representation of Simfony values with distinct / nominal types.
The Boolean type is different from the sum of two unit types,
even though the structure of both types is the same.
Similarly, a Boolean value is different from a sum-of-unit value,
even though the structure of both values is the same.