Add TypedValue
We (will) require the type of witness values to be included in the
witness JSON file, to make sure that value encodings such as "1" are
unambiguous. We (will) also check that the provided witness values fit
the type declared inside the Simfony program. In other contexts, it
might also be useful to know the type of a Simfony value.

I chose to create separate Value and TypedValue structs. This means we
first construct a Value and then we try to combine it with a type,
potentially failing. There are multiple reasons for this choice:

1. Constructing values alongside types increases code complexity
2. Simplicity values don't know their types either
   (We should first make Simplicity values aware of their type
    and then switch over to Simfony values)
3. Decoding a type first and then decoding a string as a value of this
   type always produces a well-typed value, provided the string encoding
   is valid