Get span from AsRef<Span> implementors
Most types in the parsing module have a span. Implement AsRef<Span> to
access this span in a generic manner. Span is Copy, so it can be cheaply
deferenced. Implement a conversion from AsRef<Span> to Span using
dereferencing. This enables us to write shorter code:

result.with_span(self)

instead of

result.with_span(self.span)

Span is Copy so it cheap to dereference it.