Allow empty arrays
Allow empty array types, array values and parseable array expressions.
The empty array is structurally equivalent to the empty tuple, which is
the unit. Empty arrays are not so useful, but allowing them removes one
possibly confusing error path for Simfony developers.

Note that we haven't implemented empty array patterns yet, so the "array
must be nonempty" error still exists.