Keep low-level compiler error message
The current Simfony compiler constructs Simplicity programs largely
without type-checking. This means, the construction easily fails for
various reasons. For debugging, we need the entire Simplicity error
message instead of a high-level Simfony message. We will add type
checking to the compiler so we can forget about the Simplicity details.