Add jets
Assign a Simfony source type and target type to each jet.
Compare these types with the types from rust-simplicity for structural
equality. This ensures that the Simfony "ABI" is compatible with
Simplicity's.

There is no iterator or list of all jets from rust-simplicity. Rust
enums don't provide this functionality out of the box, so I hardcoded an
array of all jets here. It might be useful to upstream the list,
potentially generating it using the Haskell-Rust code generator.