Add builtin type aliases
For every type alias in the Simplicity tech report (that is being used
in the signature of an actual jet), define a Simfony type alias that is
defined in the same way. Reserve these aliases in the grammar so users
cannot overwrite them and end up in `True = False` situations like in
Python.

These aliases will ultimately become proper types that need to be cast
to be converted into their underlying definition. However, type casts
are not implemented yet, so keep aliases for now.