Add AST
The abstract syntax tree is a well-typed representation of the program.
While the parse tree may contain various errors, these errors are caught
during construction of the abstract syntax tree. Different
representations for integers (decimal, bits, bytes) etc. can be unified
into a single representation (constant value). A scope keeps track of
variable types, of type aliases and of witness types. It makes sense to
separate the AST scope from the compilation scope, as they keep track of
different things. The compiler of the AST (future commit) will have
access to a well-typed program tree, so it doesn't need to track types.
In the future, we can use the abstract syntax tree for optimizing the
Simfony program further. In this commit, there are no optimizations.