Compile AST to Simplicity
`GlobalScope` no longer needs to handle type aliases, as the AST is
already well-typed. Demote GlobalScope to private and rename it to Scope
because it is only used during compilation. There is a different scope
for the generation of the AST. Make the compile methods private except
for Program, to prevent compilation with invalid scopes.