Add codegen subcrate
This crate generates Rust code which we will use to compile Simfony code
with the Rust compiler. For now, we generate Rust functions for each jet
with a dummy implementation. We can use rustdoc to generate
documentation from that.

For easier overview, we split the jets into categories, like in the tech
report. We group jets that have to do with arithmetic in the module for
arithmetic, and so on. We assert that all jets are in exactly one
category in a unit test.

In the future, we might assign jets to higher types such as `Vec` in
Rust land. For now, jets are bare functions that live inside a module.