Remove `test` example program
This program showed the usage of witness data using an old parser for
witness files. Remove the program because the parser will be removed in
a following commit.

The new witness file parser will take time to implement because of
pruning and other considerations around Simplicity witnesses.
Potentially we need to update rust-simplicity, which will take
additional time. For the moment, we will finalize Simfony programs with
an empty witness.