Move IdentParser into parse module
Keep IdentParser public in case it is needed elsewhere. We will decide
its visibility when we add the new witness parser, which will need to
access IdentParser.

Implement a custom version of PestParse for Program, which directly
takes a program string and hides the parsing details from the caller.
This is similar to ast::Program, which implements a custom version of
ast::AbstractSyntaxTree. Callers need to construct whole programs but
nothing else. Keep the low-level details private in the module. Make
PestParse private.