Remove old witness file parsing
The previous witness file parser used the Simplicity bit encoding
and was prone to error. The unit tests used a different parser that
returned Value::u32(20) for all witness expressions.

Remove this old parser and use simplicity::SimpleFinalizer with an empty
iterator over witness values. This will fail if any witness expression
is in the Simfony program. Display an error message that witness
expressions are temporarily disabled. Yes, it sucks that users can
compile a Simfony program with witness expressions and then it fails
last minute during satisfaction. However, keeping witness expressions in
the grammar enables users to produce program commitments to Simfony
programs with witness expressions. Also, we will add the new witness
parser soon, which will restore full support for witness expressions.
Add a TODO as a reminder.