feat: Add ConstructNode
Simfony used to use a reimplementation of NamedConstructNode from the
human encoding. NamedConstructNode is a private item of rust-simplicity.
It turns out, we don't need this complexity. We need a ConstructNode
where witness nodes contain WitnessName instead of the unit value NoWitness.
We cannot overwrite simplicity::ConstructNode, so we reimplement it
here. This turns out to be fewer lines of code than expected.

Rename ProgExt to CoreExt and implement the trait for all types that
implement CoreConstructible. CoreExt contains only default methods and
effectively extends CoreConstructible with additional useful methods.

This commit is messy, but it cleans up the NamedConstructNode complexity
and replaces it with a minimalist ConstructNode. It might help to
compare the entire file named.rs before the commit with the file after
the commit, instead of looking at the many individual changes.