feat: Create spans from strings
Create a span that covers an entire string. This will be used for
errors that don't have a precise location available and must assume
the entire string to be erroneous.

From<&str> for Span iterates over all lines of the string to compute
the final position. This operation is linear in the size of the string,
which is kind of expensive. However, I don't see a better way of
computing the final position. In any case, From<&str> for Span should be
used only in emergency situations when the precise location of the error
is not available.