feat: Compute typed value from const expression
Yes, I have accidentally implemented a restricted version of
const expressions that can be evaluated at compile time. Const
expressions consist of value constructors only. Function calls or match
expressions are not supported.

Each entry in the witness JSON file will contain a stringified type and
a stringified const expression that evaluates to a value of the given
type (else there is an parse error).