feat: Parse WitnessValues from JSON
We might want to move the serde logic into a separate module, or
introduce a macro. I leave this for a future commit until I better
understand serde and I know which architecture would work best for
Simfony. For the time being, witness JSON files are the only place where
we need serde.