feat: Convert ConstructNode to WitnessNode
Convert Simfony's ConstructNode to Simplicity's WitnessNode by
populating witness nodes with values from a witness map.

To ensure that witness nodes are populated with values of a fitting
type, WitnessValues::is_consistent must be called on the original
Simfony program before it is compiled to Simplicity. rust-simplicity
currently doesn't sanitize the values inside witness nodes, although it
might do so in the future.