feat: Add functions
This commit has a lot of lines, but most of them are mechanical changes.
We are changing the fundamental structure of a Simfony program:

Before, a program was a sequence of statements. Each statement was
delimited by a semicolon `;`.
Now, a program is a sequence of "items". Items are either type aliases
or functions. There is a special function called "main" which must exist
exactly once in every program and which must not take any inputs nor
produce any output. Functions have bodies which are block expressions.
Block expressions may contain, as before, statements.

This architecture is heavily inspired by Rust's grammar, where "modules"
consist of "items". Type aliases or functions are "items".
https://doc.rust-lang.org/reference/items/modules.html

To keep Simfony's grammar simple, I removed the possibility to define
type aliases inside function bodies.

To reduce the diff of this commit, I implemented only the main function.
Calls to custom functions are not allowed by the grammar.
Definitions of custom functions are ignored by the compiler.

To further reduce the diff of this commit, I ignored the example
programs unit test because all these programs need to be updated. I will
do so in the next commit.