feat: Add calls to custom functions
Compile the function body with a scope that contains exactly the
function parameters as variables (plus the previously defined aliases
and functions).

Witness expressions are not allowed in functions because calling a
function multiple times would result in two witness nodes with the same
name in the compiled Simplicity program. Witness expressions can be used
in the main function only.

An example for how function calls are implemented:

fn f(p1: t1, p2: t2, ...) { body }

fn main() {
    let out: Out = f(v1, v2, ...);
}

This code is inlined by the Simfony compiler to become:

fn main() {
    let out: Out = {
        let (p1, p2, ...): (t1, t2, ...) = (v1, v2, ...);
        body
    }
}