feat: Cast values of structurally equivalent types
Casts allow us to convert a pair (u8, u8) of two integers into one u16
integer, and so on. A cast between two Simfony types A and B is possible
if and only if the Simplicity types (aka structural types) that underly
A and B are exactly equal.

Casts in languages like Rust or C++ usually specify their target type.
Simfony casts specify their **source type**. The Simfony compiler knows
the expected output type of each expression. In the above example, the
compiler knows that it expects a `u16` output, but it doesn't know if
the cast input value `(1, 1)` is of type `(u8, u8)` or `(u16, u16)` or
`(u32, u32)`, ... . Therefore, Simfony casts have a slightly weird syntax:

let out: u16 = <(u8, u8)>::into((1, 1));

More generally, the syntax is

let out: TargetType = <SourceType>::into(source_value);

For what it's worth, this is valid Rust and the Rust code works in the
same way as the Simfony code. I think we can live with this syntax,
although it isn't perfect. Casts of the form
<TargetType>::from(source_value) would require a major overhaul of the
compiler.

Also remove `$` from the grammar to allow whitespaces in types,
such as `(u8, u8)`, as opposed to the spaceless `(u8,u8)`.