feat: Add list-fold expressions
A fold has the signature
  fold::<f, N>(list: List<E, N>, accumulator: A) -> A
Where the folded function has the signature
  f(element: E, accumulator: A) -> A

All elements of the list are processed by `f` in order. Each iteration
updates the accumulator, which is initialized when the fold is called.
The final accumulator is the output of the fold.

I wrote an alternative fold where `f` has readonly access to a context
which doesn't change across iterations. For brevity, I omitted this
version here. Contextual information can be passed inside the accumulator.

The above fold syntax has several advantages:
1. We avoid introducing function(pointer)s as a new type. This would be
   the case if we wrote fold::<N>(list, initial_acc, f). The "generic
   arguments" inside the diamond brackets are essentially lawless
   country where we can put non-values such as `f`.
2. The compiler knows the exact type of the parameters `list, initial_acc`
   to the fold. This means we can supply literal arguments such as
   `list![1, 2, 3]` (which could a list with any upper limit ≥ 4) or
   `42` (which could be any unsigned integer with at least 8 bits).
   Rust would require type annotations in these cases.

We translate folds into Simplicity in the next commit.