doc: Remove outdated syntax
One of the things we learned this year is that no documentation is
better than bad, misleading or false documentation. The documented
Simfony syntax is from three months ago. I will write a short rustbook
for the new syntax, but this will take time. Until then, people should
read the official rustbook on one hand and compare with the Simfony
example programs on the other hand.

https://doc.rust-lang.org/stable/book/

Also remove links to the removed files.