feat: Switch to right-associativity
Simfony tuples and arrays are translated into binary trees of Simplicity
products. We use BTreeSlice for this conversion. Switch BTreeSlice to
produce right-associative binary trees.

This resolves many tuple parameters in the signature of jets:
  fn jet(a: A, bc: (B, C))
becomes
  fn jet(a: A, b: C, c: C)

Most Simplicity jets have right-associative signatures. We have to
introduce tuple parameters for jets such as bip_0340_verify which are
left-associative. We will add special handling for these jets in a
following commit.