feat: Add panic! expression
panic! is the Simfony-encoding of a Simplicity fail node. Because panic!
never returns, it fits any output type we want. The never type `!`
doesn't exist in the Simfony type system, but at the moment we don't
seem to need it.

Simfony's panic! doesn't take any error message as argument because
Simfony doesn't support strings.

The fail nodes will be pruned out of the compiled Simfony program.