feat: Extend PairBuilder
`case` is the only Core Simplicity combinator that introduces sums in
the source type of expressions. If we ensure that our expressions
contain no `case`, then we can always pair them without error.
`comp` copies the source type of its first argument so the second
argument is unrestricted and may contain `case`.

Also move all calls of the PairBuilder constructor into its impl block,
to ensure that invariants are retained.