lockfile: reduce dependency versions to ones that MSRV can handle
I recently updated my local CI to run `cargo test --doc` in addition to
the existing tests. Inexplicably, this is causing it to try to bulid
more dependencies than it previously was, and it turns out that some
dependencies don't build on MSRV.

We have committed a lockfile in this repo, which is inadvisable for a
library, but since we did it, my local CI setup tries to use it. So
update it to something that works.