fix: keywords merge with identifiers or constants
`letx:u32=1` is not valid Rust, nor should it be valid Simfony.
`let(x,)=(1,)` is valid Rust, however, so we cannot simply ban
whitespaces.

I created an atomic rule for each keyword with the !ASCII_ALPHANUMERIC
predicate. This predicate consumes no input, but it asserts that a
keyword is not directly followed by a character that would "merge" with
the keyword.