test: Add fuzz regressions
The slow unit takes 20 seconds on my hardware and is thus ignored. The
parsing itself is slow. I will look into the grammar and see if I can
optimize it. There is a grammar debugger, but it stack-overflows if I
run it on this example :) We might need to switch to a different parsing
library to ultimately speed up the unit, but I will hold off on such a
massive change until more important work has been done.