feat: Add for-while expressions
A for-while has the signature
  for_while::<f, N>(accumulator: A, context: C) -> Either<B, A>
Where the looped function has the signature
  f(accumulator: A, context: C, counter: u{N}) -> Either<B, A>

The loop executes `f` at most `2^N` times, where `N` is a power of two.
The loop returns the first `B` value that `f` returns, or it returns
the final accumulator if `f` never returns a `B` value.

Unlike list folds, for-while loops frequently access outside information
from a readonly context. This is why the context is included in the
signature of `f`.

Including `f` as a generic parameter to `for_while` follows the example
of `fold`, which avoids function pointer values.

Simfony's syntax hides the maximum number `2^N` of loop iterations from
the  developer, but I think developers are smart enough to infer that 8
bits mean 2^8 = 256 iterations, and so on. Double powers of two are also
conceptually hard to understand, which is another reason to avoid them.

The translation to Simplicity happens in the next commit.