refactor: Move Bits | Bytes -> mod str
Rename Bits to Binary and Bytes to Hexadecimal, and make both contain a
plain Arc<str>. During parsing, check that binary / hexadecimal
characters are present, but don't do any semantic checks. Keep the code
simple. Directly convert the binary / hexadecimal strings to UIntValue
in the value module, using the code that previously lived in the parse
module.

Lots of moved lines of code.