feat: Impl Eq + Hash for types with Span
Spans are pointers into the original source file and they are useful
only for error reporting. They do not influence the function of the
Simfony tree and should thus not influence equality of the tree.

A trivial solution is to make all spans equal to one another, but this
seems wrong. Instead, I wrote a macro that implements PartialEq, Eq and
std::hash::Hash using selected members of a given type. I used this
macro for all types that include a Span and selected all non-Span
members.

The implementation of PartialEq, Eq and std::hash::Hash is still
recursive, as for #[derive(PartialEq, Eq, Hash)]. I plan to implement
equality iteratively, but this will probably include a Merkle root that
I stick into every struct. I will hold off on this change until a later
PR. For now, we have an equality that ignores Span, which is what I
wanted.