chore: Update rust-sim to f6e7ecfc43852064780bba6f019b6532d572f166
We have to introduce the inference context into all methods that
construct Simplicity expressions. This is quite verbose and I might
write an abstraction that hides the context in a later PR. For now, I
simply add the context where it is needed.

We have to be careful that all compile::Scope instances use the same
inference context. To this end, I introduce the Scope::new method that
is called once to create the main scope, and the Scope::child method
that is called when creating function scopes.

Also handle the new jets. Finally we can get rid of the hardcoded array
of all jets.