feat: Type-check Simfony values
Because we decided to type-check every Simplicity value, we also need to
type-check every Simfony value. This produces quite a large diff,
although most changes are mechanical. The implementations of
From<Option<Self>> and From<Either<Self, Self>> no longer work for Value
because additional type information is needed.

TypedValue is subsumed by Value. The implementation of Arbitrary for
Value needs to be rewritten. Typed values allow us to generate arbitrary
well-formed values, which was previously not the case. In the
implementation of Arbitrary, first a random type is generated and then a
random value of this type. There will be a function such as
`Value::arbitrary_of_type`. The details depend on how arbitrary values
are going to be used, and they are not used at the moment, so I simply
removed the implementation of Arbitrary for Value here.

To reduce the diff, the update of rust-simplicity happens in the next
commit. For now, the Simplicity values are kept typeness.