feat: Add example program test harness
Build a unit test for an example program with custom witness data and
with a custom transaction. For now, only the lock time and sequence
number can be changed, but these are the values that we introspect the
most, at the moment. Add a method for printing the sighash, in order to
generate signatures. Add a method for printing the program encoding, in
order to export the test vector to rust-simplicity.

Add #[allow(dead_code)] to the lock time and sequence methods, because
we don't use them at the moment. I will add more examples in the
following commits, which will make use of these methods.