test: Fuzz reconstruction of values
This fuzz target is 10x slower than our other targets. I suspect that
Simfony values such as [u8; 256] are cheap for the fuzzer to generate,
but they correspond to deeply nested trees of products in the Simplicity
world, which slows down the target. Maybe this is just how it is; maybe
we can be smarter about this. A saving grace is that we almost never
convert Simplicity values to Simfony values. The only use case right now
is to display debug information from the Simplicity Bit Machine.