feat: Reconstruct Simplicity values
We split a Simplicity value alongside a Simfony type in order to
reconstruct the resulting parts as a Simfony value. There is a lot of
machinery for this:

We have to undo the folding operations that we did when creating
the Simplicity type. This, we implemented in the previous commit.

We have to iterate over the Simplicity value in terms of a Simfony type.
The Destructor from this commit fulfills this role.

Putting together the destructed parts is easy, similar to how Values are
constructed from const expressions.