feat: Add debug expression
dbg!(x) behaves in Simfony like in Rust: It is a NOP that prints its
input value as a side effect. The Simfony compiler adds a debug symbol
for this purpose.

I split FallibleCall from DebugValue in the debug symbol API because
FallibleCall is supposed to be used in error messages, while DebugValue
is used for logging on the side. I don't want to handle dbg! as a case
in an error message, even though dbg! is infallible.