refactor: Represent compiled Simfony program w/o witness
We can compile a Simfony program to Simplicity without knowing the
witness data. As of now, the witness data is added directly to the
Simplicity target code.

This commit creates a new type CompiledProgram that represents the
compiled, but not-yet-satisfied, Simfony program. There are methods
to upgrade CompiledProgram to the existing SatisfiedProgram.

This commit also makes the members inside SatisfiedProgram private, as
these members are prone to change in the near future.