chore: Change look of witness expressions
witness("a") -> witness::A

The new witness expressions look like constants from the namespace
"witness". In the following commits, we will add this very namespace as
an optional part of a Simfony program. This will enable hardcoding
witness data inside the Program, for applications such as the web IDE.

Minimal code changes. Had to update all occurrences of witness
expressions inside example programs.