refactor: Make serde optional
Introduce the `serde` feature which enables parsing of witness JSON
files. In the future, witness data will also be able to be provided via
Simfony code. In this case, we don't need serde. The web IDE will not
need serde and will profit from a smaller WASM size by not including
serde.