refactor: Intergation tests
common/test.rs: The TestCase struct holds all the data that is necessary
to run an integration test. Test cases can be constructed via a builder
pattern. At the moment, there is a runtime error if necessary data is
missing. This could be changed in the future via marker types, but for
simplicity sake we don't do so in this commit.

common/util.rs: Functions for computing Simfony witness data. At the
moment, we use insecure 32-bit secure keys inside our Simfony programs.
We will upgrade to secure keys once we are able to parameterize values
such as public keys that are used inside Simfony programs.

spend_utxos.rs: The actual integration tests, which are executed in
three phases:
1) All tests create a UTXO
2) Enough blocks are mined so the lock time of each test is satisfied
3) All tests try to spend their UTXO
This pattern is possible thanks to the monotonicity of time locks,
meaning that if a time lock is satisfied at a certain block height,
it will remain satisfied in future blocks.