fix: Debug symbol CMRs differ by architecture
The CMR used in debug symbols was a hash of the literal span of the
affected function call expression. It turns out that the spans from the
PEST library differ on x86 and WASM. This resulted in entirely different
debug symbols and entirely different root CMRs of programs with debug
symbols.

To fix this, we assign each span an index. Instead of hashing the span,
we hash its index. Because the PEST parser maintains identity and order
of spans on all architectures, the resulting debug symbols are the same
on all architectures.