refactor: Generalize module parsing
Allow for different kinds of modules that each define a list of
constants. Each module has a name, such as the witness module.

Currently the only module is `mod witness { ... }`. In the following
commits, we will add a `mod param { ... }` module for defining
commitment-time values such as public keys.

Most changes in this commit are renamings or moved code.