feat: Parameterize Simfony programs
Many Simfony programs differ only in hardcoded values such as public
keys. This commit introduces parameters of the form `param::PUBLIC_KEY`
which can be used as variables that resolve to a Simfony value at
commitment time.

Simfony programs with parameters are templates which can be instantiated
as concrete Simfony programs by providing arguments for the template's
parameters.

CompiledProgram::new(s, Arguments::default()) can be used to parse and
compile a Simfony program without parameters. However, I expect most
programs to contain parameters. It remains an open question if
parameters are a good abstraction for this use case. If they turn out as
not useful, we can remove them.

I had to adapt the fuzz tests to handle arguments. Interestingly, we can
use simfony::ArbitraryOfType to create a random list of arguments for a
given list of parameters. Very nice!