refactor: WitnessTypes and WitnessValues
This commit introduces a bunch of changes to the map types WitnessTypes
and WitnessValues. Instead of splitting these changes into a dozen
commits, I combined them all into one commit. The goal of these changes
is to make WitnessTypes and WitnessValues more ergonomic to use.

Move WitnessTypes into witness.rs.

Use HashMap instead of BTreeMap for better asymptotics. I used to
require a sorted list of keys, but that is no longer the case.

Make both WitnessTypes and WitnessValues immutable. Instead of building
a mutation API for these maps, we can simply construct a HashMap with
its rich API. All values of HashMap<WitnessName, ResolvedType> are valid
for WitnessTypes and similarly for WitnessValues.

Wrap the inner map in an Arc<T> for cheap cloning. This means that we
can pass owned WitnessValues, which simplifies some code.

Wrap the implementation of WitnessTypes and of WitnessValues inside
macros, so we can reuse the code for other kinds of maps. I will add
Parameters and Arguments, which will use these macros.