chore: Print module in order of keys
simfony::Arguments and simfony::WitnessValues can be printed as
`mod param { ... }` and `mod witness { ... }`, respectively.
We almost never need this, but when we do, it is helpful to print the
module items `const NAME: TYPE = Value` in order of `NAME`.

Because simfony::Arguments and simfony::WitnessValues internally use
HashMap, their keys are not sorted. Manually sorting the keys is
computationally expensive, but we rarely need this code. Printing modules
is basically only required for the web IDE and it is not required for the
Simfony compiler itself. The Simfony compiler benefits from the armotized
complexity of the HashMap operations.