chore: Display byte arrays as hex
The implementation of fmt::Display for Value loops over each inner value
in (verbose) pre-order. This makes it hard to display a larger value in
a custom way while hiding its inner values.

The updated implementation tries to parse a Simfony array as a non-empty
byte array. If this succeeds, the bytes are displayed as hex and a flag
is set to true. The flag hides the inner values of the array, which have
already been printed. At the end, the flag is set to false again.