fix: Parse type aliases
I forgot to enforce the semicolon at the end of each type alias. This
meant that type aliases only parsed without semicolon, which is not the
intended behavior.

This commit adds a unit test to assert that type aliases can be parsed
as expected. Adding semicolons also makes one fuzz regression obsolete.