Add Non interactive siganture example
Had to refactor some of the dummy_env stuff. If I try to include
change output in the default env, everything breaks and I did not want
to spend time trying to fix all of tests. Instead, we create a new
default with fee just for this one test case.