clippy: replace flat_map with filter_map
BasePattern::as_identifier returns an Option; when iterating we can
filter out the Nones with `filter_map`. We were instead using
`flat_map`, which is much more general: it turns the Options into
iterators (which will yield 0 or 1 items) and then yields everything
from the iterators.

This is potentially slower and more complicated to read.